Fix CCL compatibility of the generator
SBCL and CCL disagree in the handling of dotfiles in their pathname
code: SBCL puts the entire file name into the pathname name, while CCL
makes the name nil and puts the file name without the leading dot as the
pathname type.

The fix is to handle the possibility of nil pathname name.