Add define-immutable-structure macro
It's pretty common for the generator code to define structures with all
slots being read-only, so add a macro that makes it more concise and
that automatically defines a trivia pattern for destructuring the
structure.

All applicable structure definitions have been changed to use that
macro, and the easy destructuring was used to improve readability in a
few places.