Enhance Pygments server protocol to handle multiline strings
This allows the generator and the server to directly send the source and
highlighted HTSL forms over the pipe, instead of using temporary files
as a crutch.

This change makes generate-everything calls about 20% faster on my
machine if /tmp is a tmpfs, and up to 80% faster if /tmp is on the disk.