Use a better-scaling root font-size formula on up-to-date browsers
The new formula gives smaller font sizes in small viewports, allowing
more text to fit; and larger ones in large viewports, decreasing the
amount of screen space wasted for side margins.

The factor of 0.6625, or 53/80, is chosen specifically to have the old
formula and the new one intersect at viewports whose smaller dimension
is 1080 pixels.

The old formula remains in place for older browsers that don't support
the CSS max function.