Decrease the base font size a bit
Some users reported that they don't like it when the text is too big, so
this change reins in the viewport-relative factor a bit.

Additionally, the viewport-relative factor now contributes nothing on
screens ≤ 800 px in minor dimension: on screens that small, mostly
phones and tablets, the browser default font size is plenty enough. This
creates a discontinuity in the font size as a function of viewport size,
but it's not a huge deal.