Refactor the Pygments cache
The Pygments cache has been moved to the greenspun.pygments package.
It now only exposes nextGeneration and highlightCode methods, so that
the previous dance with get and put is not necessary anymore, and in
fact, the HTSL converter no longer needs to talk to the Pygments server
connection pool directly anymore, it only needs the Pygments cache.