Relicense all code to AGPL 3 or later
Copyright, just like other kinds of "intellectual" "property", is a form
of theft, but since the system is what it is, I may as well make this
code as unappealing as possible to the biggest thieves — corporations —
as they run away scared whenever the very name "AGPL" is uttered.

Actual people aren't likely to be affected by the added requirement of
having to share the source code of their modifications if the blog
generator is somehow provides as a "service".