Simplify ImmutableList
Structural sharing of sublists is no longer present; there's now a
dedicated simple builder for immutable lists instead of the messy
bolted-on mechanic of freezing ArrayList.

Ideally the builder would just be a trivial subclass of ArrayList the
only adds the freeze method, but the required fields are unfortunately
inaccessible.