Bring the Pygments server protocol closer to how Java side works
Since the Java side already keeps track of the current CSS class, make
that a part of the protocol, as runs of tokens with the same CSS classes
are very common.

Additionally, add an explicit bare line feed token type, so that lone
line feeds don't have to be encoded as multiline strings.