Replace one last use of print in the Pygments server
Somehow this one last call to print was missed in
e5f778eea279a5d2986445437eadf90d31bce4fa. The reason it went undetected
for so long is that when the Pygments server is invoked normally, the
text sink is the standard output; only when used interactively with a
different text sink, perhaps when debugging or experimenting, the text
sink can be different.