Make the Pygments server protocol more regular
This change introduces clear separation of commands and simple strings,
so that it's possible to reliably distinguish a simple token ":error"
from actual Pygments server errors.

Additionally, the distinction between multiline strings that end in a
newline and ones that don't is now gone: all multiline strings now
behave like old ":mn" strings, that is, lines are separated with line
feeds, with no implicit line feed at the end.